src/requirements.txt
bcrypt==4.2.0
Flask==3.0.3
mysql-connector-python==9.0.0
requests==2.32.3
cryptography==43.0.0
exchangelib==5.0.1
sentry-sdk[flask]==2.13.0
pydantic
scikit-learn
numpy
python-docx
pypdf